# fix performance issue in wallet with offers