Skip parsing (and computing) offers just to get to the involved coins, instead use the DB field for coins_of_interest. This fixes a serious performance issue in wallets containing offers.

@Quexington please confirm that this field is in fact the same as what you get from `get_involved_coins()` on an offer.

| operation | before | after | after / before |
| -- | -- | -- | -- |
| chia wallet show | 376s | 0.75s | 0.2% |

Before this change, the wallet profile for `chia wallet show` in the test wallet looked like this:
![chia-hotspot-main](https://user-images.githubusercontent.com/661450/187076069-80f4a0e6-97f6-4706-b325-2e1285d4f97b.png)